refactor: Separate Store interface and StoreImpl
- Split Store interface and its implementation StoreImpl for separation of concerns
